+++ src/main/java/com/example/demo/base/config/Config.java
@@ -1,0 +1,40 @@
+package com.example.demo.base.config;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import jakarta.persistence.EntityManager;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
+import org.springframework.web.client.RestTemplate;
+
+import java.util.concurrent.Executor;
+
+@Configuration
+public class Config {
+
+    @Bean
+    public JPAQueryFactory jpaQueryFactory(EntityManager em) {
+        return new JPAQueryFactory(em);
+    }
+
+    @Bean
+    public RestTemplate restTemplate() {
+        return new RestTemplate();
+    }
+
+
+    @Bean
+    public Executor asyncMailThreadPool() {
+        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
+        executor.setMaxPoolSize(4);
+        executor.setKeepAliveSeconds(30);
+        executor.setCorePoolSize(4);
+        executor.setThreadNamePrefix("mail-");
+        executor.setQueueCapacity(30);
+        return executor;
+    }
+
+
+
+
+}
