+++ src/main/java/com/example/demo/base/config/RedisConfig.java
@@ -1,4 +1,5 @@
-package com.example.demo.base;
+package com.example.demo.base.config;
+
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -6,31 +7,9 @@ import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
 import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.data.redis.serializer.StringRedisSerializer;
-import org.springframework.scheduling.concurrent.ThreadPoolTaskExecutor;
-import org.springframework.web.client.RestTemplate;
-
-import java.util.concurrent.Executor;
 
 @Configuration
-public class Config {
-
-
-    @Bean
-    public Executor asyncMailThreadPool() {
-        ThreadPoolTaskExecutor executor = new ThreadPoolTaskExecutor();
-        executor.setMaxPoolSize(4);
-        executor.setKeepAliveSeconds(30);
-        executor.setCorePoolSize(4);
-        executor.setThreadNamePrefix("mail-");
-        executor.setQueueCapacity(30);
-        return executor;
-    }
-
-
-    @Bean
-    public RestTemplate restTemplate() {
-        return new RestTemplate();
-    }
+public class RedisConfig {
 
     @Bean
     public RedisConnectionFactory redisConnectionFactory() {
@@ -53,5 +32,4 @@ public class Config {
 
         return  redisTemplate;
     }
-
 }
