+++ src/main/java/com/example/demo/boundedContext/member/repository/CustomMemberRepository.java
@@ -1,0 +1,8 @@
+package com.example.demo.boundedContext.member.repository;
+
+import com.example.demo.boundedContext.member.entity.Member;
+
+import java.util.List;
+
+public interface CustomMemberRepository {
+}
