+++ src/main/java/com/example/demo/boundedContext/member/repository/MemberRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface MemberRepository extends JpaRepository<Member, Long> {
+public interface MemberRepository extends JpaRepository<Member, Long> ,CustomMemberRepository{
     Optional<Member> findByUsername(String username);
     Optional<Member> findByIdAndDeleteDateIsNull(Long id);
     Optional<Member> findByUsernameAndDeleteDateIsNull(String username);
