+++ src/main/java/com/example/demo/boundedContext/member/repository/MemberRepositoryImpl.java
@@ -1,0 +1,15 @@
+package com.example.demo.boundedContext.member.repository;
+
+import com.querydsl.jpa.impl.JPAQueryFactory;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
+import org.springframework.transaction.annotation.Transactional;
+
+@RequiredArgsConstructor
+@Transactional
+@Repository
+public class MemberRepositoryImpl implements CustomMemberRepository {
+
+    private final JPAQueryFactory jpaQueryFactory;
+
+}
